producer_bluesky/requirements.txt
pandas
kafka-python
requests
psycopg2-binary